+++ store/core/src/main/java/org/apache/carbondata/store/devapi/DataScanner.java
@@ -17,15 +17,18 @@
 
 package org.apache.carbondata.store.devapi;
 
+import java.io.Serializable;
 import java.util.Iterator;
 
 import org.apache.carbondata.common.annotations.InterfaceAudience;
 import org.apache.carbondata.common.annotations.InterfaceStability;
+import org.apache.carbondata.core.scan.expression.Expression;
+import org.apache.carbondata.sdk.store.descriptor.TableIdentifier;
 import org.apache.carbondata.sdk.store.exception.CarbonException;
 
 @InterfaceAudience.Developer("Integration")
 @InterfaceStability.Unstable
-public interface DataScanner<T> {
+public interface DataScanner<T> extends Serializable {
 
   /**
    * Perform a scan in a distributed compute framework like Spark, Presto, etc.
